change code and fix index of rolController
