+++ app/Http/Controllers/gestion_rol/RolController.php
@@ -5,107 +5,106 @@ namespace App\Http\Controllers\gestion_rol;
 use App\Http\Controllers\Controller;
 use App\Models\Rol;
 use App\Permission\PermissionConst;
+use App\Util\QueryUtil;
 use Illuminate\Http\Request;
 use Spatie\Permission\Contracts\Permission;
 use Spatie\Permission\Models\Role;
 
 class RolController extends Controller
 {
-    public function __construct()
-    {
-        // $this->middleware('permission:' . PermissionConst::GESTION_ROLES);
+  public function __construct()
+  {
+    // $this->middleware('permission:' . PermissionConst::GESTION_ROLES);
+  }
+
+  /**
+   * Display a listing of the resource.
+   * @param  \Illuminate\Http\Request  $request
+   * @return \Illuminate\Http\Response
+   */
+  public function index(Request $request)
+  {
+    $nombre = $request->input('name');
+    $idCompany = $request->input('idCompany');
+
+    $roles = Rol::with("company");
+
+    if ($nombre) {
+      $roles->where('name', '=', $nombre);
     }
 
-    /**
-     * Display a listing of the resource.
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function index(Request  $request)
-    {
-        $nombre = $request->input('name');
-        $company = $request->input('company');
-
-
-        $roles = Rol::with("company");
-
-        if ($nombre) {
-            $roles->where('name', '=', $nombre);
-        }
-
-        if ($company) {
-            $roles->whereHas('company', function ($q) use ($company) {
-                $q->where('id', '=', $company);
-            });
-        }
-
-
-        return response()->json($roles->get());
+    if ($idCompany) {
+      $roles->whereHas('company', function ($q) use ($idCompany) {
+        $q->where('id', '=', $idCompany);
+      });
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        $data = $request->all();
-        $rol = new Rol($data);
-        $rol->save();
+    return response()->json($roles->get());
+  }
 
-        return response()->json($rol, 201);
-    }
+  /**
+   * Store a newly created resource in storage.
+   *
+   * @param  \Illuminate\Http\Request  $request
+   * @return \Illuminate\Http\Response
+   */
+  public function store(Request $request)
+  {
+    $data = $request->all();
+    $rol = new Rol($data);
+    $rol->save();
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show(int $id)
-    {
-        $rol = Rol::find($id);
+    return response()->json($rol, 201);
+  }
 
-        return response()->json($rol);
-    }
+  /**
+   * Display the specified resource.
+   *
+   * @param  int $id
+   * @return \Illuminate\Http\Response
+   */
+  public function show(int $id)
+  {
+    $rol = Rol::find($id);
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, int $id)
-    {
-        $data = $request->all();
-        $rol = Rol::findOrFail($id);
-        $rol->fill($data);
-        $rol->save();
+    return response()->json($rol);
+  }
 
-        return response()->json($rol);
-    }
+  /**
+   * Update the specified resource in storage.
+   *
+   * @param  \Illuminate\Http\Request  $request
+   * @param  int $id
+   * @return \Illuminate\Http\Response
+   */
+  public function update(Request $request, int $id)
+  {
+    $data = $request->all();
+    $rol = Rol::findOrFail($id);
+    $rol->fill($data);
+    $rol->save();
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(int $id)
-    {
-        $rol = Rol::findOrFail($id);
-        $rol->delete();
+    return response()->json($rol);
+  }
 
-        return response()->json([], 204);
-    }
+  /**
+   * Remove the specified resource from storage.
+   *
+   * @param  int $id
+   * @return \Illuminate\Http\Response
+   */
+  public function destroy(int $id)
+  {
+    $rol = Rol::findOrFail($id);
+    $rol->delete();
 
-    // public function index()
-    // {
-    //     $roles = Rol::with('company')->get();
+    return response()->json([], 204);
+  }
 
-    //     return response()->json($roles);
-    // }
+  // public function index()
+  // {
+  //     $roles = Rol::with('company')->get();
+
+  //     return response()->json($roles);
+  // }
 }
